add plot and scatter
